requirements.txt
requests
python-dotenv
PyPDF2
groq
